WeedIoTNew/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
multidict==6.7.0
narwhals==2.7.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.0
protobuf==6.32.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.14.1
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.22.0
